requirements.txt
Django>=5.0,<5.1
gunicorn
python-decouple
requests
whitenoise
django-allauth[socialaccount]
django-allauth-ui
django-widget-tweaks
slippers
stripe